AI: Prevent switch-loop due to Sucker Punch.
Sucker Punch must not be labeled a a good supereffective move. Because of how it works, AI will switch into a "good" mon (has Sucker Punch), then the score calculation can just return a bad score for it due to circumstances, and the mon that just got sent out will get switched back due to AllBadMoves, repeating the cycle indefinitely.