Dealt with lock_confuse cases and simplified skyDropTarget values.
When a target is targeted by Sky Drop while in the middle of Petal Dance/Outrage/other moves that cause them to be confused eventually, the target immediately becomes confused "due to fatigue" once they are released, regardless of how they come to be released.
Also too ghoulslash's advice and made the cleared value for skyDropTargets to be 0xFF, which made a lot of conditional statements a lot shorter.