prevent inf. conf. being set when conf. is prevented.
if a pokemon with own tempo uses baton pass, it would pass it the inf. conf. bit to the next mon. if that mon became confused then, it would be confused infinitely.
So we need to make sure this bit is never set for a pokemon with own tempo.
Since misty terrain prevents the confusion also, this applies aswell.