Fix veneering via removing -fno-toplevel-reorder from most files
The reason -fno-toplevel-reorder existed is because some code relied on ordering of functions as they were copied to a buffer. There is probably a way to do this by changing the C code, but this works for now. This allows the compiler to reorder functions to minimize veneering.

I only added it for the files that require it, plus agb_flash.c

agb_flash.c happens to not be affected by this, but it has a portion where it copies code to a buffer so I added it just to be safe.