Have shaymin revert to land forme when frozen
This is a little obscure, but a fun fact is that if shaymin ends up frozen, it reverts back to land form until you use the gracadia flower again, which is after the battle.

Note: I have yet to figure out how to prevent the gracidiea flower from working if shaymin is still frozen when you try to use said flower. Also does not account for the frozen status being set outside of battle, if that is a thing.