-fno-toplevel-reorder is only needed for m4a, agb_flash, and librfu_intr
-fno-toplevel-reorder prevents the compiler from moving around functions and ensures the memory layout of the assembler of each function matches the source code order.

Relying on a particular order of functions is UB, but some code like the one in m4a, relies on this for the sake of using the start of other functions to determine the size a function would be for the sake of copying code to memory. This has been tested so the game boots and works.